Server_2/src/ai_python_service/requirements.txt
fastapi
uvicorn[standard]
torch
soundfile
numpy
TTS
openai-whisper
transformers
accelerate
safetensors
